feat: controller分层，添加传感器数据，date转换方法封装
