+++ backend/src/main/java/com/fall/smarthouse/service/ISensorService.java
@@ -1,5 +1,10 @@
 package com.fall.smarthouse.service;
 
+import com.fall.smarthouse.model.Sensor;
+
+import java.sql.Timestamp;
+import java.text.ParseException;
+
 /**
  * @author FAll
  * @date 2022/12/2 17:19
@@ -7,4 +12,11 @@ package com.fall.smarthouse.service;
  */
 public interface ISensorService {
 
+    /**
+     * @description: 向传感器表中添加数据
+     * @author xiaoQe
+     * @date 2022/12/5 15:19
+     * @version 1.0
+     */
+    public boolean insertToSensor(String time,Double gas,Double smog,Double temperature,Double humidity,Double shake) throws ParseException;
 }
