+++ backend/src/main/java/com/fall/smarthouse/service/impl/SensorServiceImpl.java
@@ -1,7 +1,16 @@
 package com.fall.smarthouse.service.impl;
 
+import com.fall.smarthouse.mapper.SensorMapper;
+import com.fall.smarthouse.model.Sensor;
 import com.fall.smarthouse.service.ISensorService;
+import com.fall.smarthouse.util.DateConverter;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.sql.Date;
+import java.sql.Timestamp;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
 
 /**
  * @author FAll
@@ -10,4 +19,22 @@ import org.springframework.stereotype.Service;
 @Service
 public class SensorServiceImpl implements ISensorService {
 
+    @Autowired
+    SensorMapper sensorMapper;
+
+
+    @Override
+    public boolean insertToSensor(String time, Double gas, Double smog, Double temperature,
+                                  Double humidity, Double shake) throws ParseException {
+        String dateString = DateConverter.LongToDateString(Long.parseLong(time));
+
+        SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+        Date date = new Date(sdf.parse(dateString).getTime());
+        Sensor sensor = new Sensor(date, gas, smog, temperature, humidity, shake);
+        Integer affectRows = sensorMapper.insertToSensor(sensor);
+        if(affectRows == 0){
+            return false;
+        }
+        return true;
+    }
 }
