+++ backend/src/test/java/com/fall/smarthouse/BackendApplicationTests.java
@@ -4,6 +4,8 @@ import com.fall.smarthouse.mapper.ElectricMapper;
 import com.fall.smarthouse.mapper.SensorMapper;
 import com.fall.smarthouse.model.ElectricAppliance;
 import com.fall.smarthouse.model.Sensor;
+import com.fall.smarthouse.service.IElectricApplianceService;
+import com.fall.smarthouse.service.ISensorService;
 import com.fall.smarthouse.service.impl.ElectricApplianceServiceImpl;
 import com.fall.smarthouse.util.JWTUtil;
 import org.junit.jupiter.api.Test;
@@ -11,7 +13,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
 import java.sql.Timestamp;
+import java.text.SimpleDateFormat;
 import java.util.Calendar;
+import java.util.Date;
 import java.util.Scanner;
 
 @SpringBootTest
@@ -24,10 +28,21 @@ class BackendApplicationTests {
     ElectricMapper electricMapper;
 
     @Autowired
-    ElectricApplianceServiceImpl electricApplianceService;
+    IElectricApplianceService electricApplianceService;
+
+    @Autowired
+    ISensorService sensorService;
+
+    @Test
+    void testSensorService(){
+//        System.out.println(sensorService.insertToSensor(new Sensor(new Timestamp(Calendar.getInstance().getTimeInMillis()),
+//                1.0,2.0,3.3,5.0,6.7)));
+    }
+
 
     @Test
     void testUpdateElectricAppliance(){
+        System.out.println(System.currentTimeMillis());
         System.out.println(electricApplianceService.setWarnLight(0));
     }
 
@@ -70,6 +85,14 @@ class BackendApplicationTests {
         boolean ret = JWTUtil.isNeedUpdate("Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzUxMiJ9.eyJzdWIiOiLlvKDkuIkiLCJleHAiOjE2NzAxNDE3MDl9.X6bAVUGkMypLQZOgpDHuG99nE4Nx85JSVmjZjnD0ThXC-vbm1jDjUpYhsVikhpT2nJ11XNnkdxltzknzxn3jiw");
         System.out.println(ret);
     }
+
+    @Test
+    void testSimpleDateFormat() {
+        SimpleDateFormat sdf=new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+        String format = sdf.format(new Date(System.currentTimeMillis()));
+        System.out.println(format);
+    }
+
 
 
 }
